requirements.txt
streamlit>=1.28.0
moviepy>=2.1.0
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-chroma>=0.1.0
sentence-transformers>=2.2.0
FlagEmbedding>=1.2.0
chromadb>=0.4.0
numpy>=1.26.0,<2.2.0